Modified Locale variable names
Locale variable names changed to AditionalLanguage.
Added Title variable to be used on search services as aditional show
name.
Modified migration class.
