--- src/NzbDrone.Core/Datastore/Migration/080_add_locale_clean_title_to_series.cs
@@ -1,44 +1,0 @@
-﻿using System.Data;
-using FluentMigrator;
-using NzbDrone.Core.Datastore.Migration.Framework;
-
-namespace NzbDrone.Core.Datastore.Migration
-{
-    [Migration(80)]
-    public class add_locale_clean_title_to_series : NzbDroneMigrationBase
-    {
-        protected override void MainDbUpgrade()
-        {
-            Alter.Table("Series").AddColumn("LocaleCleanTitle").AsString().Nullable();
-
-            Execute.WithConnection(SetLocaleCleanTitle);
-        }
-
-        private void SetLocaleCleanTitle(IDbConnection conn, IDbTransaction tran)
-        {
-            using (IDbCommand getSeriesCmd = conn.CreateCommand())
-            {
-                getSeriesCmd.Transaction = tran;
-                getSeriesCmd.CommandText = @"SELECT Id, CleanTitle FROM Series";
-                using (IDataReader seriesReader = getSeriesCmd.ExecuteReader())
-                {
-                    while (seriesReader.Read())
-                    {
-                        var id = seriesReader.GetInt32(0);
-                        var cleanTitle = seriesReader.GetString(1);
-
-                        using (IDbCommand updateCmd = conn.CreateCommand())
-                        {
-                            updateCmd.Transaction = tran;
-                            updateCmd.CommandText = "UPDATE Series SET LocaleCleanTitle = ? WHERE Id = ?";
-                            updateCmd.AddParameter(cleanTitle);
-                            updateCmd.AddParameter(id);
-
-                            updateCmd.ExecuteNonQuery();
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
