+++ src/NzbDrone.Core/Tv/Series.cs
@@ -21,9 +21,10 @@ namespace NzbDrone.Core.Tv
         public int TvRageId { get; set; }
         public string ImdbId { get; set; }
         public string Title { get; set; }
+        public string AditionalLanguageTitle { get; set; }
         public string CleanTitle { get; set; }
         public string SortTitle { get; set; }
-        public string LocaleCleanTitle { get; set; }
+        public string AditionalLanguageCleanTitle { get; set; }
         public SeriesStatusType Status { get; set; }
         public string Overview { get; set; }
         public string AirTime { get; set; }
